build: remove setting libdirs in premake4.lua
"libdirs { libdir }" in the main premake4.lua file causes Debug_x32
directory to be added for the libdirs for all builds. I think that's a
bug in premake, as at the same place setting, say, "targetdir" works
fine.

However, setting "libdir" to "libdirs" is not needed at all, so let's remove it.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@iki.fi>